requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
black==23.7.0
blinker==1.6.2
cachetools==5.3.1
cffi==1.15.1
chardet==5.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
distlib==0.3.6
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pyproject-api==1.5.3
pyrsistent==0.19.3
pytest==7.4.0
pytest-black==0.3.12
pytest-flake8==1.1.1
pytest-flask==1.2.0
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.17
toml==0.10.2
tox==4.6.4
typing_extensions==4.7.0
virtualenv==20.23.1
Werkzeug==2.3.6
